+++ nano_ktrans/utils/__init__.py
@@ -1,7 +1,14 @@
 from .context import get_context, set_context, reset_context
 from .expert_selection import generate_gpu_experts_masks, uniform_gpu_experts_masks, profile_expert_activation
+from .expert_runtime_state import (
+    ExpertMigrationOp,
+    ExpertResidency,
+    ExpertResidencyPlan,
+    LayerExpertState,
+)
 
 __all__ = [
     "get_context", "set_context", "reset_context",
     "generate_gpu_experts_masks", "uniform_gpu_experts_masks", "profile_expert_activation",
+    "ExpertMigrationOp", "ExpertResidency", "ExpertResidencyPlan", "LayerExpertState",
 ]